Update description of github releases in docs
Resolves issue 3433. I also ended up rearranging docs/conf.py a bit.
