+++ docs/conf.py
@@ -19,12 +19,6 @@ panels_add_bootstrap_css = False
 
 project = "Pyodide"
 copyright = "2019-2022, Pyodide contributors and Mozilla"
-pyodide_version = "0.22.0"
-
-if ".dev" in pyodide_version or os.environ.get("READTHEDOCS_VERSION") == "latest":
-    CDN_URL = "https://cdn.jsdelivr.net/pyodide/dev/full/"
-else:
-    CDN_URL = f"https://cdn.jsdelivr.net/pyodide/v{pyodide_version}/full/"
 
 # -- General configuration ---------------------------------------------------
 
@@ -133,20 +127,14 @@ htmlhelp_basename = "Pyodidedoc"
 # A list of files that should not be packed into the epub file.
 epub_exclude_files = ["search.html"]
 
-
-def delete_attrs(cls):
-    """Prevent attributes of a class or module from being documented.
-
-    The top level documentation comment of the class or module will still be
-    rendered.
-    """
-    for name in dir(cls):
-        if not name.startswith("_"):
-            try:
-                delattr(cls, name)
-            except Exception:
-                pass
-
+base_dir = Path(__file__).resolve().parent.parent
+extra_sys_path_dirs = [
+    str(base_dir),
+    str(base_dir / "pyodide-build"),
+    str(base_dir / "docs/sphinx_pyodide"),
+    str(base_dir / "src/py"),
+    str(base_dir / "packages/micropip/src"),
+]
 
 # Try not to cause side effects if we are imported incidentally.
 
@@ -158,6 +146,31 @@ except ImportError:
     IN_SPHINX = False
 
 IN_READTHEDOCS = "READTHEDOCS" in os.environ
+
+if IN_SPHINX or IN_READTHEDOCS:
+    sys.path = extra_sys_path_dirs + sys.path
+    import builtins
+
+    from sphinx_pyodide.util import docs_argspec
+
+    # override docs_argspec, _pyodide.docs_argspec will read this value back.
+    # Must do this before importing pyodide!
+    setattr(builtins, "--docs_argspec--", docs_argspec)
+
+import pyodide
+
+# The full version, including alpha/beta/rc tags.
+release = version = pyodide.__version__
+
+if ".dev" in version or os.environ.get("READTHEDOCS_VERSION") == "latest":
+    CDN_URL = "https://cdn.jsdelivr.net/pyodide/dev/full/"
+else:
+    CDN_URL = f"https://cdn.jsdelivr.net/pyodide/v{version}/full/"
+
+html_title = f"Version {version}"
+
+global_replacements = {"{{PYODIDE_CDN_URL}}": CDN_URL, "{{VERSION}}": version}
+
 
 if IN_READTHEDOCS:
     env = {"PYODIDE_BASE_URL": CDN_URL}
@@ -185,75 +198,14 @@ if IN_READTHEDOCS:
     console_path.write_text("".join(console_html))
 
 
-def docs_argspec(argspec: str) -> Any:
-    """Decorator to override the argument spec of the function that is
-    rendered in the docs.
-
-    If the documentation finds a __wrapped__ attribute, it will use that to
-    render the argspec instead of the function itself. Assign an appropriate
-    fake function to __wrapped__ with the desired argspec.
-    """
-
-    def dec(func):
-        d = func.__globals__
-        TEMP_NAME = "_xxxffff___"
-        assert TEMP_NAME not in d
-        code = dedent(
-            f"""\
-            def {TEMP_NAME}{argspec}:
-                pass
-            """
-        )
-        # exec in func.__globals__ context so that type hints don't case NameErrors.
-        exec(code, d)
-        f = d.pop(TEMP_NAME)
-
-        # # Run update_wrapper but keep f's annotations and don't set
-        # # f.__wrapped__ (would cause an infinite loop!)
-        annotations = f.__annotations__
-        update_wrapper(f, func)
-        f.__annotations__ = annotations
-        del f.__wrapped__
-
-        # Set wrapper to be our fake function
-        func.__wrapped__ = f
-
-        return func
-
-    return dec
-
-
 if IN_SPHINX:
-    base_dir = Path(__file__).resolve().parent.parent
-    path_dirs = [
-        str(base_dir),
-        str(base_dir / "pyodide-build"),
-        str(base_dir / "docs/sphinx_pyodide"),
-        str(base_dir / "src/py"),
-        str(base_dir / "packages/micropip/src"),
-    ]
-    sys.path = path_dirs + sys.path
-
     from sphinx.domains.javascript import JavaScriptDomain, JSXRefRole
 
     JavaScriptDomain.roles["class"] = JSXRefRole()
 
-    import builtins
-    from functools import update_wrapper
-    from textwrap import dedent
-
-    # override docs_argspec, _pyodide.docs_argspec will read this value back.
-    # Must do this before importing pyodide!
-    setattr(builtins, "--docs_argspec--", docs_argspec)
-
-    import pyodide
     from pyodide.ffi import JsProxy
 
     del JsProxy.__new__
-
-    # The full version, including alpha/beta/rc tags.
-    release = version = pyodide.__version__
-    html_title = f"Version {version}"
 
     shutil.copy("../src/core/pyproxy.ts", "../src/js/pyproxy.gen.ts")
     shutil.copy("../src/core/error_handling.ts", "../src/js/error_handling.gen.ts")
@@ -275,6 +227,8 @@ if IN_SPHINX:
 
     # Prevent API docs for webloop methods: they are the same as for base event loop
     # and it clutters api docs too much
+    from sphinx_pyodide.util import delete_attrs
+
     import pyodide.console
     import pyodide.webloop
 
@@ -293,8 +247,6 @@ def globalReplace(app, docname, source):
         result = result.replace(key, app.config.global_replacements[key])
     source[0] = result
 
-
-global_replacements = {"{{PYODIDE_CDN_URL}}": CDN_URL}
 
 always_document_param_types = True
